requirements.txt
annotated-types==0.7.0flask
openai
pdfplumber
pytesseract
python-docx
gunicorn

anyio==4.8.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2flask
openai
pdfplumber
pytesseract
python-docx
gunicorn

distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
lxml==5.3.1
openai==1.65.5
packaging==24.2
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pypdfium2==4.30.1
pytesseract==0.3.13
python-docx==1.1.2
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
